requirements.txt
appnope==0.1.0
attrs==19.3.0
autograd==1.3
backcall==0.1.0
bleach==3.1.0
certifi==2020.6.20
chardet==3.0.4
Click==7.0
cmdstanpy==0.4.0
convertdate==2.2.0
cycler==0.10.0
Cython==0.29.14
decorator==4.4.1
defusedxml==0.6.0
dill==0.3.1.1
entrypoints==0.3
ephem==3.7.7.0
future==0.18.2
gspread==3.3.0
holidays==0.9.11
httplib2==0.17.0
idna==2.9
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
jedi==0.16.0
Jinja2==2.10.3
joblib==0.14.0
json5==0.8.5
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.2
jupyterlab==1.1.4
jupyterlab-server==1.0.0
kiwisolver==1.1.0
Lifetimes==0.11.1
LunarCalendar==0.0.9
lunardate==0.2.0
MarkupSafe==1.1.1
matplotlib==3.1.3
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.1
numpy==1.18.1
oauth2client==4.1.3
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.1
patsy==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.3
psycopg2==2.7.7
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.5.2
PyMeeus==0.3.6
pyparsing==2.4.6
pyrsistent==0.15.7
pystan==2.19.1.1
python-dateutil==2.8.1
python-dotenv==0.10.5
pytz==2019.3
pyzmq==18.1.1
requests==2.23.0
rsa==4.0
scikit-learn==0.21.3
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
setuptools-git==1.2
simplegeneric==0.8.1
six==1.14.0
statsmodels==0.11.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
xlrd==1.2.0
zipp==2.2.0